Put sockets in a diff file in a diff directory
